move boottest/* into boottest/utils/
