+++ boottest/utils/hasher.py
